Moves indicator from behavior ophys -> behavior metadata
